requirements.txt
cassandra-driver>=3.28.0
python-dotenv>=1.0.0
PyYAML>=6.0
